Skip DAO tests if neo4j is not available
