+++ tests/test_dao.py
@@ -1,30 +1,14 @@
 import os
-from typing import Generator, Optional
+from typing import  Optional
 
 import pytest
-from neo4j import Driver, GraphDatabase
-from yarl import URL
+from neo4j import Driver
 
 from neo4j_python_migrations.dao import MigrationDAO
 from neo4j_python_migrations.migration import Migration, MigrationType
+from .conftest import can_connect_to_neo4j, username
 
-username = os.environ["NEO4J_MIGRATIONS_USER"]
-password = os.environ["NEO4J_MIGRATIONS_PASS"]
-host = os.environ["NEO4J_MIGRATIONS_HOST"]
-port = int(os.environ["NEO4J_MIGRATIONS_PORT"])
-scheme = os.environ["NEO4J_MIGRATIONS_SCHEME"]
-
-
-@pytest.fixture
-def neo4j_driver() -> Generator[Driver, None, None]:
-    with GraphDatabase.driver(
-        str(URL.build(scheme=scheme, host=host, port=port)),
-        auth=(username, password),
-    ) as driver:
-        yield driver
-        with driver.session() as session:
-            session.run("MATCH (n) DETACH DELETE n")
-
+pytestmark = pytest.mark.skipif(not can_connect_to_neo4j(), reason="Neo4j is not available")
 
 def test_create_baseline(neo4j_driver: Driver) -> None:
     dao = MigrationDAO(neo4j_driver)
